Added new branch and merged steps for SauceDemo
