+++ WspólnySpecFlowProject/WspólnySpecFlowProject.csproj
@@ -18,4 +18,8 @@
     <PackageReference Include="FluentAssertions" Version="6.2.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Pages\" />
+  </ItemGroup>
+
 </Project>
